// server/enums/educationLevelValue.ts
enum EducationLevelValue {
  PRIMARY_SCHOOL = 'PRIMARY_SCHOOL',
  SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS = 'SECONDARY_SCHOOL_LEFT_BEFORE_TAKING_EXAMS',
  SECONDARY_SCHOOL_TOOK_EXAMS = 'SECONDARY_SCHOOL_TOOK_EXAMS',
  FURTHER_EDUCATION_COLLEGE = 'FURTHER_EDUCATION_COLLEGE',
  UNDERGRADUATE_DEGREE_AT_UNIVERSITY = 'UNDERGRADUATE_DEGREE_AT_UNIVERSITY',
  POSTGRADUATE_DEGREE_AT_UNIVERSITY = 'POSTGRADUATE_DEGREE_AT_UNIVERSITY',
  NOT_SURE = 'NOT_SURE',
}

export default EducationLevelValue